added more testing

+++ tests/load_points_from_obj/script.py
@@ -2,7 +2,7 @@
 def remove_texture(point):
     out = []
     for a in point:
-        out.append(float(a.split('//')[0]))
+        out.append(int(a.split('//')[0]))
 
     return tuple(out)
